refactor(core): split LogFetcher.swift by responsibility (#3005)
598-line file had four distinct concerns; extract three:

- LogFetcher+Models.swift — StepLogResult, UnzipResult, ZipExtractor
- JobNameSanitisation.swift — sanitizeJobNameForZIP (gh/gha-compatible)
- LogFetcher+ZIPExtraction.swift — unzipLogs / unzipLogsTyped

LogFetcher.swift keeps only the fetcher core (415 lines). No behaviour
change.
