+++ Sources/RunBotCore/Platform/LogFetcher.swift
@@ -4,60 +4,6 @@ import Foundation
 import GitHubClient
 import os
 
-// MARK: - Filesystem path constants
-
-/// Absolute path to the system `unzip` binary, always present on macOS.
-private let unzipBinaryPath = "/usr/bin/unzip" // NOSONAR — fixed OS path
-
-// MARK: - StepLogResult
-
-/// Typed result from step log extraction.
-/// Every case is distinct — wrong content can never silently look like correct content.
-public enum StepLogResult: Equatable, Sendable {
-    /// ZIP file found and matched by {sanitisedJobName}/{stepNumber}_*.txt.
-    case slice(content: String)
-    /// ZIP contained no per-step files for this job — flat blob used instead.
-    /// StepLogView MUST render a visible degradation notice for this case.
-    case flatBlobFallback(content: String)
-    /// Step produced no output (ZIP file present but empty, or step not found).
-    case syntheticEmpty(stepName: String, reason: String)
-    /// Network failure, ZIP parse failure, or subprocess failure.
-    case fetchFailed(reason: String)
-
-    /// The log text to display, or nil if there is nothing to show.
-    public var text: String? {
-        switch self {
-        case .slice(let content), .flatBlobFallback(let content): return content
-        case .syntheticEmpty, .fetchFailed: return nil
-        }
-    }
-
-    /// True when this result represents a step that was intentionally skipped by the runner
-    /// (as opposed to a fetch failure or a step that ran but produced no output).
-    /// StepLogView uses this to render a neutral “Step skipped” headline instead of
-    /// the error-toned “No output recorded” headline.
-    public var isSkipped: Bool {
-        guard case .syntheticEmpty(_, let reason) = self else { return false }
-        return reason.hasPrefix("This step was skipped")
-    }
-}
-
-// MARK: - UnzipResult
-
-/// Typed result from `unzipLogs(_:)` — distinguishes subprocess failure from empty archive.
-public enum UnzipResult: Sendable {
-    /// ZIP extracted successfully; contains (relative path, text) pairs for every `.txt` file found.
-    case success([(name: String, text: String)])
-    /// `/usr/bin/unzip` exited with a non-zero status. The associated value is the raw exit code.
-    case processFailed(exitCode: Int32)
-    /// A filesystem operation failed (directory creation, file write, or enumeration).
-    case ioError
-}
-
-/// A closure that extracts a ZIP archive (`Data`) into named text file entries.
-/// The default implementation spawns `/usr/bin/unzip`; tests inject a stub.
-public typealias ZipExtractor = @Sendable (Data) async -> UnzipResult
-
 // MARK: - LogFetcher
 
 /// Injectable fetcher for GitHub Actions job and workflow-run logs.
@@ -466,133 +412,4 @@ public struct LogFetcher: Sendable {
                 : "This job had step log files, but none matched step \(step.number). Available files for this job: \(available). Tried prefix: '\(prefix)' (sanitised job name: '\(sanitised)')."
         )
     }
-}
-
-// MARK: - Job name sanitisation for ZIP lookup
-
-/// Sanitises a GitHub job name for use as a ZIP folder prefix.
-///
-/// Applies three rules in order, matching the logic in `getJobNameForLogFilename` in the
-/// `gh` CLI ([`logs.go`](https://github.com/cli/cli/blob/trunk/pkg/cmd/run/view/logs.go))
-/// and the C# server's `GetJobNameForLogFilename`:
-/// 1. Strip characters invalid in Windows file paths — the C# server (which writes the ZIP)
-///    strips `/`, `:`, `"`, `*`, `?`, `<`, `>`, `|`. Not stripping these causes a prefix
-///    mismatch when a job name contains any of them and the server has already removed them
-///    from the ZIP folder name.
-/// 2. Truncate to **90 UTF-16 code units** — the GitHub server (C#) truncates using
-///    `String.Substring(0, 90)` which counts UTF-16 `Char` objects. Swift `.count` counts
-///    Unicode scalars, not UTF-16 code units — emoji and CJK characters truncate at a
-///    different offset without this step.
-/// 3. Trim leading/trailing whitespace/newlines — the C# server calls `.Trim()` after truncation
-///    (and the `gh` CLI does `strings.TrimSpace`). Truncation can expose trailing
-///    whitespace that was previously interior.
-func sanitizeJobNameForZIP(_ name: String) -> String {
-    // Strip the full set of characters the C# server considers invalid in a Windows path.
-    // The `gh` CLI only strips `/` and `:` explicitly, but the server strips all of these.
-    // Stripping only `/` and `:` causes silent misses when job names contain `"`, `*`, etc.
-    let stripped = name
-        .replacingOccurrences(of: "/", with: "")
-        .replacingOccurrences(of: ":", with: "")
-        .replacingOccurrences(of: "\"", with: "")
-        .replacingOccurrences(of: "*", with: "")
-        .replacingOccurrences(of: "?", with: "")
-        .replacingOccurrences(of: "<", with: "")
-        .replacingOccurrences(of: ">", with: "")
-        .replacingOccurrences(of: "|", with: "")
-    guard stripped.utf16.count > 90 else { return stripped.trimmingCharacters(in: .whitespacesAndNewlines) }
-    var units = Array(stripped.utf16.prefix(90))
-    // Guard against splitting a surrogate pair at the 90-unit boundary.
-    // Swift strings are always well-formed, so surrogates only appear as
-    // high+low pairs. The only reachable split is a high surrogate at position
-    // 89 (0xD800–0xDBFF) whose low partner was at 90 and is now dropped —
-    // remove the dangling high to keep the output valid.
-    // A lone low surrogate (0xDC00–0xDFFF) at the boundary cannot arise from a
-    // well-formed Swift String: the high surrogate at 88 would still be present,
-    // keeping the pair intact. Malformed UTF-16 fed via String(decoding:as:UTF16.self)
-    // is normalised to U+FFFD before we ever see it, so no guard is needed.
-    if let last = units.last, (0xD800...0xDBFF).contains(last) {
-        units.removeLast()
-    }
-    // Mirror the C# server's `.Trim()` call (and the gh CLI's `strings.TrimSpace`) which
-    // both trim the result after truncation. Truncation can expose trailing whitespace
-    // that was previously interior.
-    return String(decoding: units, as: UTF16.self).trimmingCharacters(in: .whitespacesAndNewlines)
-}
-
-// MARK: - ZIP extraction (uses /usr/bin/unzip — always available on macOS)
-
-/// Extracts all `.txt` files from a ZIP blob and returns `(name, text)` pairs.
-///
-/// Writes the ZIP to a unique temporary directory, runs `/usr/bin/unzip -q` via
-/// `ProcessRunner.runAsync`, then enumerates the output directory for `.txt` files.
-/// The temporary directory is always removed on return via `defer`.
-///
-/// - Parameter zipData: Raw ZIP archive bytes as returned by the GitHub logs API.
-/// - Returns: An array of `(name, text)` tuples where `name` is the archive-relative
-///   path without the `.txt` extension (e.g. `"release/1_Build"` for `release/1_Build.txt`)
-///   and `text` is the file content. Returns `[]` if the write, unzip, or enumeration
-///   step fails.
-func unzipLogs(_ zipData: Data) async -> [(name: String, text: String)] {
-    switch await unzipLogsTyped(zipData) {
-    case .success(let files): return files
-    case .processFailed, .ioError: return []
-    }
-}
-
-/// Typed variant of `unzipLogs` that distinguishes subprocess failure from an empty archive.
-///
-/// Returns `.processFailed(exitCode:)` when `/usr/bin/unzip` exits non-zero (so
-/// `fetchStepLog` can map it to `.fetchFailed(reason: "Invalid repository scope: expected owner/repo.")` rather than `.syntheticEmpty`).
-/// Returns `.ioError` when the temp directory or ZIP file cannot be created.
-func unzipLogsTyped(_ zipData: Data) async -> UnzipResult {
-    let fileManager = FileManager.default
-    let tmp = fileManager.temporaryDirectory.appendingPathComponent(UUID().uuidString)
-    let zipFile = tmp.appendingPathComponent("logs.zip")
-    defer { try? fileManager.removeItem(at: tmp) }
-    do {
-        try fileManager.createDirectory(at: tmp, withIntermediateDirectories: true)
-        try zipData.write(to: zipFile)
-    } catch { return .ioError }
-    let result = await ProcessRunner.runAsync(
-        executableURL: URL(fileURLWithPath: unzipBinaryPath),
-        arguments: ["-q", zipFile.path, "-d", tmp.path]
-    )
-    // Exit code 0 = success. Exit code 1 = success with warnings (e.g. extra bytes
-    // prepended to the ZIP — which GitHub's log API occasionally produces). Only
-    // exit code 2 and above indicate a genuine extraction failure.
-    guard result.exitCode <= 1 else { return .processFailed(exitCode: result.exitCode) }
-    guard let enumerator = fileManager.enumerator(at: tmp, includingPropertiesForKeys: nil) else { return .ioError }
-    // Keep only .txt files and exclude the archive itself.
-    // `logs.zip` is written into `tmp` before extraction so the enumerator sees it;
-    // the `.pathExtension == "txt"` check is the primary guard ("zip" ≠ "txt"), and
-    // the explicit URL comparison is belt-and-suspenders against a future rename of
-    // the temp file to a .txt name.
-    let zipFileResolved = zipFile.resolvingSymlinksInPath()
-    let txtURLs = enumerator.compactMap { $0 as? URL }.filter {
-        $0.pathExtension == "txt"
-        && $0.resolvingSymlinksInPath() != zipFileResolved
-        // macOS's zip tool injects __MACOSX/ AppleDouble metadata entries into every
-        // archive it creates. These are never valid step log files and would pollute
-        // allFiles, skew stepCount, and add noise to diagnostic logs.
-        && !$0.path.contains("/__MACOSX/")
-    }
-    var results: [(name: String, text: String)] = []
-    // Resolve symlinks on the tmp prefix so that the /tmp → /private/tmp alias
-    // on macOS does not cause the prefix-stripping below to silently no-op,
-    // leaving absolute paths (e.g. "/privaterelease/2_Checkout") in the names.
-    let tmpResolved = tmp.resolvingSymlinksInPath().path
-    for url in txtURLs {
-        // Strip the tmp directory prefix to get the archive-relative path, then
-        // drop the .txt extension. Using NSString.deletingPathExtension rather than
-        // URL(fileURLWithPath:).deletingPathExtension().path avoids the leading-slash
-        // bug: URL(fileURLWithPath:) treats its argument as absolute, prepending "/"
-        // to relative strings and breaking the hasPrefix("jobName/") lookup downstream.
-        let resolved = url.resolvingSymlinksInPath().path
-        let relative = resolved.replacingOccurrences(of: tmpResolved + "/", with: "")
-        let name = (relative as NSString).deletingPathExtension
-        if let text = try? String(contentsOf: url, encoding: .utf8) {
-            results.append((name: name, text: text))
-        }
-    }
-    return .success(results)
 }
